OrbitalsTheme enabled for web UI.

+++ orbitals-render/src/commonMain/kotlin/org/beatonma/orbitals/render/options/ColorOptions.kt
@@ -21,11 +21,10 @@ data class ColorOptions(
     ),
     val foregroundAlpha: Float = 1f,
 ) {
-    fun colorFor(body: Body): Color =
-        Color(bodies.random().colors.random())
+    fun colorFor(body: Body): Color = Color(bodies.random().colors.random())
 }
 
-enum class ObjectColors(val colors: Array<Int>) {
+enum class ObjectColors(val colors: Array<ULong>) {
     Greyscale(MaterialGrey),
     Red(MaterialRed),
     Orange(MaterialOrange),
